# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Aqar Data Studio — Professional Real Estate Data Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties with pagination, filtering, and sorting
      parameters:
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: regionId
          in: query
          schema: { type: string }
        - name: typeId
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: minArea
          in: query
          schema: { type: number }
        - name: maxArea
          in: query
          schema: { type: number }
        - name: finishing
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: sortBy
          in: query
          schema: { type: string, default: createdAt }
        - name: sortDir
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Paginated property list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"

    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /properties/import:
    post:
      operationId: importProperties
      tags: [properties]
      summary: Import properties from parsed file data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /properties/export:
    post:
      operationId: exportProperties
      tags: [properties]
      summary: Export properties in the specified format
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Export file (binary or text)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /properties/bulk:
    post:
      operationId: bulkProperties
      tags: [properties]
      summary: Bulk operations on multiple properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkOperationInput"
      responses:
        "200":
          description: Bulk operation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkOperationResult"
  /properties/parse-text:
    post:
      operationId: parsePropertyText
      tags: [properties]
      summary: Smart property extraction from free Arabic/English text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseTextInput"
      responses:
        "200":
          description: Extracted property fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseTextResult"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a single property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /properties/{id}/duplicate:
    post:
      operationId: duplicateProperty
      tags: [properties]
      summary: Duplicate a property (creates a copy with a new code)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Duplicated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/history:
    get:
      operationId: getPropertyHistory
      tags: [properties]
      summary: Get version history for a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropertyHistoryEntry"

  # ─── REGIONS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeleteResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        id: { type: string }

    # ── Property ────────────────────────────────────────────────────────────
    Property:
      type: object
      required: [id, code, title, createdAt]
      properties:
        id: { type: string }
        code: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        area: { type: number }
        beds: { type: number }
        baths: { type: number }
        floors: { type: number }
        floor: { type: number }
        finishing: { type: string }
        view: { type: string }
        typeId: { type: string }
        regionId: { type: string }
        category: { type: string }
        status: { type: string }
        featured: { type: boolean }
        agentType: { type: string }
        images: { type: array, items: { type: string } }
        videoUrl: { type: string }
        externalUrl: { type: string }
        mapsUrl: { type: string }
        createdAt: { type: string }
        unitType: { type: string }
        subArea: { type: string }
        layout: { type: string }
        master: { type: string }
        elevator: { type: string }
        floorText: { type: string }
        location: { type: string }
        source: { type: ["string", "null"] }
        sourcePhones: { type: array, items: { type: string } }
        sourceEmail: { type: ["string", "null"] }
        sourceLocation: { type: ["string", "null"] }
        sourceNotes: { type: ["string", "null"] }
        views: { type: number }
        coverPriority: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: ["string", "null"] }
        regionName: { type: ["string", "null"] }
        typeName: { type: ["string", "null"] }
    PropertyInput:
      type: object
      required: [code]
      properties:
        code: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        area: { type: number }
        beds: { type: number }
        baths: { type: number }
        floors: { type: number }
        floor: { type: number }
        finishing: { type: string }
        view: { type: string }
        typeId: { type: string }
        regionId: { type: string }
        category: { type: string }
        status: { type: string }
        featured: { type: boolean }
        agentType: { type: string }
        images: { type: array, items: { type: string } }
        videoUrl: { type: string }
        externalUrl: { type: string }
        mapsUrl: { type: string }
        unitType: { type: string }
        subArea: { type: string }
        layout: { type: string }
        master: { type: string }
        elevator: { type: string }
        floorText: { type: string }
        location: { type: string }
        source: { type: string }
        sourcePhones: { type: array, items: { type: string } }
        sourceEmail: { type: string }
        sourceLocation: { type: string }
        sourceNotes: { type: string }
        coverPriority: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: string }
    PropertyUpdate:
      type: object
      properties:
        code: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        area: { type: number }
        beds: { type: number }
        baths: { type: number }
        floors: { type: number }
        floor: { type: number }
        finishing: { type: string }
        view: { type: string }
        typeId: { type: string }
        regionId: { type: string }
        category: { type: string }
        status: { type: string }
        featured: { type: boolean }
        agentType: { type: string }
        images: { type: array, items: { type: string } }
        videoUrl: { type: string }
        externalUrl: { type: string }
        mapsUrl: { type: string }
        unitType: { type: string }
        subArea: { type: string }
        layout: { type: string }
        master: { type: string }
        elevator: { type: string }
        floorText: { type: string }
        location: { type: string }
        source: { type: string }
        sourcePhones: { type: array, items: { type: string } }
        sourceEmail: { type: string }
        sourceLocation: { type: string }
        sourceNotes: { type: string }
        coverPriority: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: string }
    PropertyListResponse:
      type: object
      required: [data, total, page, limit, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
        totalPages: { type: number }
    PropertyHistoryEntry:
      type: object
      required: [id, propertyId, action, changedAt]
      properties:
        id: { type: string }
        propertyId: { type: string }
        action: { type: string }
        changedAt: { type: string }
        changedBy: { type: ["string", "null"] }
        snapshot: { type: object }
        diff: { type: object }

    # ── Import / Export ────────────────────────────────────────────────────
    ImportedRow:
      type: object
      required: [code]
      properties:
        code: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        area: { type: number }
        beds: { type: number }
        baths: { type: number }
        floors: { type: number }
        floor: { type: number }
        finishing: { type: string }
        view: { type: string }
        typeId: { type: string }
        regionId: { type: string }
        category: { type: string }
        status: { type: string }
        featured: { type: boolean }
        agentType: { type: string }
        unitType: { type: string }
        subArea: { type: string }
        layout: { type: string }
        master: { type: string }
        elevator: { type: string }
        floorText: { type: string }
        location: { type: string }
        source: { type: string }
        videoUrl: { type: string }
        externalUrl: { type: string }
        mapsUrl: { type: string }
    ImportInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ImportedRow"
        mode:
          type: string
          enum: [insert, update, merge, skip]
          default: merge
        dryRun:
          type: boolean
          default: false
    ImportResult:
      type: object
      required: [added, updated, skipped, errors]
      properties:
        added: { type: number }
        updated: { type: number }
        skipped: { type: number }
        errors: { type: number }
        dryRun: { type: boolean }
        details:
          type: array
          items:
            type: object
    ExportInput:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [csv, excel, json, txt, pdf]
        columns:
          type: array
          items: { type: string }
        filters:
          type: object
        sortBy: { type: string }
        sortDir: { type: string, enum: [asc, desc] }

    # ── Bulk Operations ────────────────────────────────────────────────────
    BulkOperationInput:
      type: object
      required: [operation, ids]
      properties:
        operation:
          type: string
          enum: [delete, update, archive, activate, feature, unfeature]
        ids:
          type: array
          items: { type: string }
          minItems: 1
        updates:
          type: object
    BulkOperationResult:
      type: object
      required: [success, affected]
      properties:
        success: { type: boolean }
        affected: { type: number }
        errors:
          type: array
          items: { type: string }

    # ── Smart Parsing ─────────────────────────────────────────────────────
    ParseTextInput:
      type: object
      required: [text]
      properties:
        text: { type: string }
    ParseTextResult:
      type: object
      properties:
        area: { type: ["integer", "null"] }
        beds: { type: ["integer", "null"] }
        baths: { type: ["integer", "null"] }
        price: { type: ["integer", "null"] }
        finishing: { type: ["string", "null"] }
        view: { type: ["string", "null"] }
        regionId: { type: ["string", "null"] }
        regionName: { type: ["string", "null"] }
        subArea: { type: ["string", "null"] }
        floor: { type: ["integer", "null"] }
        floorText: { type: ["string", "null"] }
        unitType: { type: ["string", "null"] }
        layout: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        confidence: { type: number }

    # ── Region ────────────────────────────────────────────────────────────
...