# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aqar Data Studio — Professional Real Estate Data Management Platform
 * OpenAPI spec version: 1.0.0
 */

export interface PropertyUpdate {
  code?: string;
  title?: string;
  description?: string;
  price?: number;
  area?: number;
  beds?: number;
  baths?: number;
  floors?: number;
  floor?: number;
  finishing?: string;
  view?: string;
  typeId?: string;
  regionId?: string;
  category?: string;
  status?: string;
  featured?: boolean;
  agentType?: string;
  images?: string[];
  videoUrl?: string;
  externalUrl?: string;
  mapsUrl?: string;
  unitType?: string;
  subArea?: string;
  layout?: string;
  master?: string;
  elevator?: string;
  floorText?: string;
  location?: string;
  source?: string;
  sourcePhones?: string[];
  sourceEmail?: string;
  sourceLocation?: string;
  sourceNotes?: string;
  coverPriority?: string;
  tags?: string[];
  notes?: string;
}